Rename package

+++ src/yeo/main.py
@@ -5,9 +5,13 @@ from pathlib import Path
 
 
 def create_config_file():
-    # TODO: If castle.json already exists, give warning to user before
-    # reinitializing it
-    with open("castle.json", "w") as f:
+    if Path("yeo.json").exists():
+        response = input("yeo.json already exists. Overwrite? (y/n): ")
+        if response.lower() != "y":
+            print("File not written.")
+            return
+
+    with open("yeo.json", "w") as f:
         json.dump(
             {
                 "paths": [
@@ -19,10 +23,11 @@ def create_config_file():
             f,
             indent=2,
         )
+    print("yeo.json created.")
 
 
 def copy_files():
-    with open("castle.json", "r") as f:
+    with open("yeo.json", "r") as f:
         config_dict = json.load(f)
 
     config_files = config_dict["paths"]
@@ -40,11 +45,12 @@ def sync():
 
 
 def main():
-    parser = argparse.ArgumentParser(prog="castle")
+    parser = argparse.ArgumentParser(prog="yeo")
     subparsers = parser.add_subparsers(dest="command")
 
-    subparsers.add_parser("init", help="Create a castle.json file")
+    subparsers.add_parser("init", help="Create a yeo.json file")
     subparsers.add_parser("sync", help="Sync your dotfiles to the current directory")
+    subparsers.add_parser("hello", help="Sync your dotfiles to the current directory")
 
     args = parser.parse_args()
 
@@ -52,6 +58,8 @@ def main():
         create_config_file()
     elif args.command == "sync":
         copy_files()
+    elif args.command == "hello":
+        print("Hello world!")
 
 
 if __name__ == "__main__":
